move keyVault POC to another file
