+++ sdk/keyvault/keyvault/src/az_keyvault_client.c
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <az_credential.h>
 #include <az_keyvault.h>
 
 #include <_az_cfg.h>
